chore: refactor sign schema
